+++ DHLClient/APIArguments/MyDHLExpressAPIArgs/Shipment/ShipmentProofOfDeliveryAPIArgs.cs
@@ -12,7 +12,7 @@
         /// </summary>
         /// <example>9356579890</example>
         [ArgumentName("shipmentTrackingNumber")]
-        public double ShipmentTrackingNumber { get; set; }
+        public long ShipmentTrackingNumber { get; set; }
 
         /// <summary>
         /// DHL Express customer shipper account number
