+++ DHLClient/APIModels/RequestModels/Pickup/PickupShipmentDetailRequestModel.cs
@@ -67,7 +67,7 @@ namespace DHLClient
         /// </summary>
         /// <example>123456790</example>
         [JsonProperty("shipmentTrackingNumber")]
-        public double ShipmentTrackingNumber { get; set; }
+        public long ShipmentTrackingNumber { get; set; }
 
         /// <summary>
         /// The package details
