+++ DHLClient/APIModels/RequestModels/Pickup/ShipmentDetailValueAddedServicesrequestModel.cs
@@ -42,7 +42,7 @@ namespace DHLClient
         /// </summary>
         /// <example>123456790</example>
         [JsonProperty("shipmentTrackingNumber")]
-        public double? ShipmentTrackingNumber { get; set; }
+        public long? ShipmentTrackingNumber { get; set; }
 
         /// <summary>
         /// The package details
